+++ pydantic_xml/serializers/factories/__init__.py
@@ -1,1 +1,2 @@
-from . import heterogeneous, homogeneous, mapping, model, primitive, tagged_union, typed_mapping, union, wrapper
+from . import heterogeneous, homogeneous, is_instance, mapping, model, primitive, raw, tagged_union, typed_mapping
+from . import union, wrapper